pyproject.toml
# SPDX-FileCopyrightText: © 2024 The "Whiteprints template context" contributors <whiteprints@pm.me>
#
# SPDX-License-Identifier: MIT

[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "whiteprints-template-context"
description = "Interpolate variables and set Copier Jinja context"
keywords = [
    "Python",
]
license = "MIT"
license-files.paths = [
    "LICENSES/CC-BY-NC-SA-4.0.txt",
    "LICENSES/MIT.txt",
]
maintainers = [
    { name = "Romain Brault", email = "" },
]
authors = [
    { name = "Romain Brault", email = "" },
]
requires-python = ">=3.9,<4.0"
classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Console",
    "Framework :: Pytest",
    "Framework :: Sphinx",
    "Framework :: tox",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dynamic = [
    "readme",
    "version",
]
dependencies = [
    "copier-templates-extensions>=0.3",
    "jinja2-time>=0.2",
    "license-expression>=30.3.1",
    "typing-extensions>=4.12.2; python_full_version<'3.12'",
]

urls.changelog = "https://github.com/whiteprints/whiteprints-template-context/releases"
urls.discussions = "https://github.com/whiteprints/whiteprints-template-context/discussions"
urls.documentation = "https://github.com/whiteprints/whiteprints-template-context"
urls.download = "https://github.com/whiteprints/whiteprints-template-context/releases"
urls.homepage = "https://github.com/whiteprints/whiteprints-template-context"
urls.integration = "https://github.com/whiteprints/whiteprints-template-context/actions"
urls.issues = "https://github.com/whiteprints/whiteprints-template-context/issues"
urls.pypi = "https://pypi.org/project/whiteprints-template-context"
urls.repository = "https://github.com/whiteprints/whiteprints-template-context"

[tool.uv]
managed = true
dev-dependencies = [
    "scalene>=1.5.19",
    "pynvim>=0.4.3",
    "beartype>=0.18.5",
    "hypothesis>=6.110.1",
    "pylint>=3.2.6",
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "pytest-html>=4.1.1",
    "pytest-randomly>=3.15.0",
    "pytest-sugar>=1.0.0",
    "pytest-xdist>=3.6.1",
    "xdoctest>=1.2.0",
]
